+++ python_bot/__init__.py
@@ -1,0 +1,7 @@
+from .notebook import NoteBook
+from .addressbook import AddressBook
+from .command_handlers import function
+from .main import main
+
+
+__all__ = ["main", "function", "AddressBook", "NoteBook"]
